learning how to organize code in go
